Start test of read configuration and status

+++ informPConfigurator/mainwindow.h
@@ -69,8 +69,8 @@ private:
     void setStatusState(QVector<uint8_t> &configBuff);
     bool getUint16FromStringInt(uint16_t *rezConvertOut, QString inString);
     bool getInt16FromStringInt(int16_t *rezConvertOut, QString inString);
-    void setModuleStatusLineEdit(QLineEdit *statuSlineEdit, uint8_t statusIndex);
-    void setDeviceStatusLineEdit(QLineEdit *statuSlineEdit, uint8_t statusIndex);
+    void setModuleStatusLineEdit(QLineEdit *statuSlineEdit, uint16_t statusIndex);
+    void setDeviceStatusLineEdit(QLineEdit *statuSlineEdit, uint16_t statusIndex);
 
 private:
     QStringListModel   *deviceList;
